Week-04/Labs/requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.7.5
tqdm==4.67.3
Pillow==10.4.0
typing-extensions==4.13.2
